requirements.txt
GitPython
dash
plotly
pandas
packaging
pyyaml
radon
kaleido
